pipeline/pyproject.toml
[tool.poetry]
name = "pipeline"
version = "0.1.0"
description = ""
authors = ["kavisek <kavi.skhon@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pyyaml = "^6.0.1"
alembic = "^1.11.3"
jinja2 = "^3.1.2"
sqlalchemy = "^2.0.20"
psycopg2-binary = "^2.9.7"
gspread = "^5.10.0"
google-cloud-bigquery = "^3.11.4"
google-cloud-logging = "^3.6.0"
google-cloud-pubsub = "^2.18.2"
storage = "^0.0.4.3"
pandas = "^2.0.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
